Allocated a separate module, repository and viewmodel for the Tutor Search feature.
