+++ app/src/main/java/com/nailorsh/repeton/data/repositories/FakeRepetonRepository.kt
@@ -2,11 +2,9 @@ package com.nailorsh.repeton.data.repositories
 
 import com.nailorsh.repeton.data.sources.FakeChatsSource
 import com.nailorsh.repeton.data.sources.FakeLessonSource
-import com.nailorsh.repeton.data.sources.FakeTutorsSource
 import com.nailorsh.repeton.domain.repositories.RepetonRepository
 import com.nailorsh.repeton.model.Chat
 import com.nailorsh.repeton.model.Lesson
-import com.nailorsh.repeton.model.Tutor
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.withContext
@@ -14,10 +12,6 @@ import java.time.LocalDate
 import javax.inject.Inject
 
 class FakeRepetonRepository @Inject constructor() : RepetonRepository {
-    override suspend fun getTutors(): List<Tutor> = withContext(Dispatchers.IO) {
-        FakeTutorsSource.getTutorsList()
-    }
-
     override suspend fun getLessons(): List<Lesson> = withContext(Dispatchers.IO) {
 //        delay(5000)
         FakeLessonSource.loadLessons()
@@ -38,6 +32,4 @@ class FakeRepetonRepository @Inject constructor() : RepetonRepository {
         delay(2000)
         FakeLessonSource.loadLessons()[id]
     }
-
-
 }
