+++ app/src/main/java/com/nailorsh/repeton/domain/repositories/RepetonRepository.kt
@@ -2,11 +2,9 @@ package com.nailorsh.repeton.domain.repositories
 
 import com.nailorsh.repeton.model.Chat
 import com.nailorsh.repeton.model.Lesson
-import com.nailorsh.repeton.model.Tutor
 import java.time.LocalDate
 
 interface RepetonRepository {
-    suspend fun getTutors(): List<Tutor>
     suspend fun getLessons(): List<Lesson>
     suspend fun getLessons(day : LocalDate) : List<Lesson>
     suspend fun getChats(): List<Chat>
