+++ app/src/main/java/com/nailorsh/repeton/domain/repositories/TutorSearchRepository.kt
@@ -1,0 +1,7 @@
+package com.nailorsh.repeton.domain.repositories
+
+import com.nailorsh.repeton.model.Tutor
+
+interface TutorSearchRepository {
+    suspend fun getTutors(): List<Tutor>
+}
